[REFACTOR] Split dashboard page into reusable components and UI helpers
